requirements.txt
Django>=5.2.0
# Dependencias para Separador de Recibos PDF
PyPDF2==3.0.1
pdfplumber==0.10.3
reportlab==4.0.7
opencv-python==4.8.1.78
django-storages==1.14.2
celery==5.3.4
redis==5.0.1
PyMuPDF>=1.24.0
pdf2image==1.17.0
matplotlib==3.7.2
Wand==0.6.13
python-dateutil==2.8.2
Pillow>=12.0.0
